% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{auc}
\alias{auc}
\title{Calculate AUC for a set of binary outcomes and a numeric vector of
prediction values.}
\usage{
auc(gold, pred)
}
\arguments{
\item{gold}{A binary vector of outcome values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate AUC for a set of binary outcomes and a numeric vector of
prediction values.
}
