% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringFunctions.R
\name{binarize}
\alias{binarize}
\title{Binarize a numeric vector above and below a given threshold.}
\usage{
binarize(vec, threshold, direction)
}
\arguments{
\item{vec}{Vector to binarize.}

\item{threshold}{Threshold at which to binarize.}

\item{direction}{Side of threshold to set to 1. greaterThan (0>thresh>=1) or lessThan (1=<thresh<0)}
}
\description{
Binarize a numeric vector above and below a given threshold.
}
