% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrappingMetric}
\alias{bootstrappingMetric}
\title{Create an matrix of bootstrapped predictions.}
\usage{
bootstrappingMetric(goldStandardMatrix, predictionsMatrix,
  scoreFun = scoreFun, bootstrapN = bootstrapN, seed = seed,
  doParallel = F, ...)
}
\arguments{
\item{goldStandardMatrix}{A single column matrix with the gold standard predictions.}

\item{predictionsMatrix}{Columns of prediction sets, in the same order as the goldStandardMatrix.}

\item{bootstrapN}{Number of total bootstraps to perform.}

\item{seed}{Set a seed for bootstrap sampling.}

\item{doParallel}{Bootstrap in parallel. Only works on UNIX based OS. Default FALSE.}
}
\value{
An MxN matrix of bootstrapped predictions where M is the number of bootstraps performed and N is the number of prediction sets.
}
\description{
Create an matrix of bootstrapped predictions.
}
