% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{c_statistic}
\alias{c_statistic}
\title{Calculate concordance index (Harrell's C-statistic)
for a set of predicted values. .}
\usage{
c_statistic(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values
(e.g. output from survival::Surv).}

\item{pred}{A numeric vector of prediction values
(e.g. output from survival::Surv).}
}
\description{
Calculate concordance index (Harrell's C-statistic)
for a set of predicted values. .
}
