% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challengescoring.R
\docType{package}
\name{challengescoring-package}
\alias{challengescoring}
\alias{challengescoring-package}
\title{challengescoring}
\description{
This package provides an implementation of the BootLadderBoot approach to scoring computational challenges whilst avoiding test data leakage.
}
\author{
\strong{Maintainer}: Robert Allaway \email{robert.allaway@sagebionetworks.org}

}
