% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{computeBayesFactor}
\alias{computeBayesFactor}
\title{Calculate one or more Bayes factors using a bootstrapMetricMatrix.}
\usage{
computeBayesFactor(bootstrapMetricMatrix, refPredIndex, invertBayes)
}
\arguments{
\item{bootstrapMetricMatrix}{An NxM matrix where M is the number of bootstraps performed and N is the number of prediction sets (output of bootstrappingMetric function).}

\item{refPredIndex}{Column index of the reference prediction to calculate Bayes factor (e.g. best prediction).}

\item{invertBayes}{Boolean to invert Bayes factor.
if(largerIsBetter == T & currentPred<=refPred){invertBayes = T},
if(largerIsBetter == T & currentPred>refPred){invertBayes = F},
if(largerIsBetter == F & currentPred<=refPred){invertBayes = T},
if(largerIsBetter == F & currentPred>refPred){invertBayes = F}}
}
\value{
A matrix of Bayes factors.
}
\description{
Calculate one or more Bayes factors using a bootstrapMetricMatrix.
}
