% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringFunctions.R
\name{pearson}
\alias{pearson}
\title{Calculate Pearson correlation for a set of predicted values.}
\usage{
pearson(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate Pearson correlation for a set of predicted values.
}
