% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringFunctions.R
\name{rmse}
\alias{rmse}
\title{Calculate root mean squared error for a set of predicted values.}
\usage{
rmse(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate root mean squared error for a set of predicted values.
}
