% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_auc}
\alias{score_auc}
\title{Calculate the Area Under the ROC Curve}
\usage{
score_auc(gold, pred)
}
\arguments{
\item{gold}{A factor vector of gold standard values.}

\item{pred}{A factor vector of prediction values.}
}
\description{
Calculate the Area Under the ROC Curve
}
