% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_balanced_accuracy}
\alias{score_balanced_accuracy}
\title{Calculate the balanced accuracy [ i.e., ( sensitivity + specificity ) / 2 ]}
\usage{
score_balanced_accuracy(gold, pred)
}
\arguments{
\item{gold}{A factor vector of gold standard values.}

\item{pred}{A factor vector of prediction values.}
}
\description{
Calculate the balanced accuracy [ i.e., ( sensitivity + specificity ) / 2 ]
}
