% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_f1}
\alias{score_f1}
\title{Calculate F1 score}
\usage{
score_f1(gold, pred)
}
\arguments{
\item{gold}{A factor vector of gold standard values.}

\item{pred}{A factor vector of prediction values.}
}
\description{
Calculate F1 score
}
