% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_mcc}
\alias{score_mcc}
\title{Calculate Matthew's Correlation Coefficient}
\usage{
score_mcc(gold, pred)
}
\arguments{
\item{gold}{A factor vector of gold standard values.}

\item{pred}{A factor vector of prediction values.}
}
\description{
Calculate Matthew's Correlation Coefficient
}
