% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_pearson}
\alias{score_pearson}
\title{Calculate Pearson correlation for a set of predicted values.}
\usage{
score_pearson(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate Pearson correlation for a set of predicted values.
}
