% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_rmse}
\alias{score_rmse}
\title{Calculate root mean squared error for a set of predicted values.}
\usage{
score_rmse(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate root mean squared error for a set of predicted values.
}
