% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_spearman}
\alias{score_spearman}
\title{Calculate Spearman correlation for a set of predicted values.}
\usage{
score_spearman(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate Spearman correlation for a set of predicted values.
}
