% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringFunctions.R
\name{spearman}
\alias{spearman}
\title{Calculate Spearman correlation for a set of predicted values.}
\usage{
spearman(gold, pred)
}
\arguments{
\item{gold}{A numeric vector of gold standard values.}

\item{pred}{A numeric vector of prediction values.}
}
\description{
Calculate Spearman correlation for a set of predicted values.
}
