% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{validate_combined_df}
\alias{validate_combined_df}
\title{Validate Combined DF}
\usage{
validate_combined_df(
  df,
  name_columns = "prediction_name",
  pred_column = "prediction",
  val_column = "validation"
)
}
\arguments{
\item{df}{A combined prediction and validation df, typically a result of
combine_validation_prediction_dfs}

\item{name_columns}{A character vector of columns used to identify a
prediction}

\item{pred_column}{A string, the column in the prediction df that has the
prediction values}

\item{val_column}{A string, the column in the validation df that has the
validation values}
}
\description{
Validate Combined DF
}
