% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{validate_required_columns}
\alias{validate_required_columns}
\title{Validate Required Columns}
\usage{
validate_required_columns(
  prediction_df,
  name_columns = "prediction_name",
  pred_column = "prediction"
)
}
\arguments{
\item{df}{A df from a prediction file}

\item{required_columns}{A character of vector of columns the df should have}
}
\description{
Validate Required Columns
}
