% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{validate_submission}
\alias{validate_submission}
\title{Validate Submission}
\usage{
validate_submission(
  validation_df,
  prediction_df,
  name_columns = "prediction_name",
  pred_column = "prediction",
  val_column = "validation",
  prediction_validation_funs = NULL,
  combined_validation_funs = NULL
)
}
\arguments{
\item{validation_df}{A df made from the validation file}

\item{prediction_df}{A df made from the prediction file}

\item{name_columns}{A character vector of columns used to identify a
prediction}

\item{pred_column}{A string, the column in the prediction df that has the
prediction values}

\item{val_column}{A string, the column in the validation df that has the
validation values}

\item{prediction_validation_funs}{A list of functions to run on the
prediction df. These functions must take params prediction_df, name_columns,
and pred_column.}

\item{combined_validation_funs}{A list of functions to run on the
combined_df. These functions must take the param combined_df.}
}
\description{
Validate Submission
}
