% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{bootLadderBoot}
\alias{bootLadderBoot}
\title{Calculate a bootstrapped score for an initial submission or subsequent submission.}
\usage{
bootLadderBoot(predictionsPath, predictionColname, goldStandardPath,
  goldStandardColname, prevPredictionsPath = NULL, scoreFun = spearman,
  bootstrapN = 10000, reportBootstrapN = 10, bayesThreshold = 3,
  seed = 98121, largerIsBetter = TRUE, verbose = FALSE,
  doParallel = TRUE)
}
\arguments{
\item{predictionsPath}{The relative path to the current prediction csv.}

\item{predictionColname}{The name of the column in the prediction csv that contains numeric prediction values. If also using a previous prediction file, must be the same name.}

\item{goldStandardPath}{The relative path to the gold standard/test data csv.}

\item{goldStandardColname}{The name of the column in the gold standard csv that contains numeric prediction values.}

\item{prevPredictionsPath}{If a previous prediction file for this team/participant already exists, pass in the path here. Prediction colname must match.}

\item{scoreFun}{A scoring function. Default is Spearman correlation. Any function can be passed as long as it can calculate a score from two vectors (gold standard first, and prediction values second).}

\item{bootstrapN}{Number of total bootstraps to perform (default 10000).}

\item{reportBootstrapN}{Number of bootstraps to base returned score off of (default 10). The greater this value, the more accurate of a result is returned (and possibly the more the test data can be overfit).}

\item{bayesThreshold}{The threshold for reporting a score. If the bayes factor (K) of the new prediction relative to the previous submission is less than this value, no score is returned. Default 3.}

\item{seed}{Set a seed for bootstrap sampling. Default 98121.}

\item{largerIsBetter}{Set this to FALSE if a smaller scoring metric indicates better performance (e.g. root mean squared error). Default TRUE.}

\item{verbose}{Report step. Default FALSE.}

\item{doParallel}{Bootstrap in parallel. Only works on UNIX based OS. Default TRUE.}
}
\description{
Calculate a bootstrapped score for an initial submission or subsequent submission.
}
